MODINVUP-100 Tech debt
   Split delete request handling and upsert request handling into separate classes because the two processes have almost no logic in common.
