+++ src/main/java/org/folio/inventoryupdate/UpdatePlan.java
@@ -14,7 +14,6 @@ import io.vertx.ext.web.RoutingContext;
 import org.folio.inventoryupdate.entities.*;
 import org.folio.inventoryupdate.entities.InventoryRecord.Entity;
 import org.folio.inventoryupdate.entities.InventoryRecord.Transaction;
-import org.folio.inventoryupdate.instructions.ProcessingInstructionsDeletion;
 import org.folio.okapi.common.OkapiClient;
 
 import io.vertx.core.CompositeFuture;
@@ -50,24 +49,10 @@ import static org.folio.inventoryupdate.InventoryUpdateOutcome.OK;
  */
 public abstract class UpdatePlan {
 
-    // The record set to update Inventory with - either coming in with the request when creating/updating
-    // or being derived from existing records in Inventory when deleting
-    protected InventoryRecordSet updatingSet;
-    protected InventoryQuery instanceQuery;
-    // Existing Inventory records matching either an incoming record set or a set of deletion identifiers
-    protected InventoryRecordSet existingSet = null;
-    // A secondary existing set that needs updating too (currently relevant only in the context of a shared index).
-    protected InventoryRecordSet secondaryExistingSet;
-
     protected static final Logger logger = LoggerFactory.getLogger("inventory-update");
     protected boolean isDeletion = false;
 
     protected Repository repository;
-
-    protected UpdatePlan (InventoryRecordSet incomingInventoryRecordSet, InventoryQuery existingInstanceQuery) {
-        this.updatingSet = incomingInventoryRecordSet;
-        this.instanceQuery = existingInstanceQuery;
-    }
 
     protected UpdatePlan () {}
 
@@ -207,7 +192,6 @@ public abstract class UpdatePlan {
         return promise.future();
     }
 
-
     protected Future<List<InventoryUpdateOutcome>> chainSingleRecordUpserts(RoutingContext routingContext, List<JsonArray> arraysOfOneRecordSet, BiFunction<RoutingContext, JsonArray, Future<InventoryUpdateOutcome>> upsertMethod) {
         Promise<List<InventoryUpdateOutcome>> promise = Promise.promise();
         List<InventoryUpdateOutcome> outcomes = new ArrayList<>();
@@ -229,49 +213,6 @@ public abstract class UpdatePlan {
         return promise.future();
     }
 
-    // DELETION
-    /**
-     * Constructor for plan for creating or updating an Inventory record set
-     * @param existingInstanceQuery The query to use for checking if the instance already exists
-     */
-    protected UpdatePlan (InventoryQuery existingInstanceQuery) {
-        this.instanceQuery = existingInstanceQuery;
-        this.isDeletion = true;
-    }
-
-    public abstract Future<Void> planInventoryDelete(OkapiClient client, ProcessingInstructionsDeletion deleteInstructions);
-
-    protected Future<InventoryRecordSet> lookupExistingRecordSet (OkapiClient okapiClient, InventoryQuery instanceQuery) {
-        Promise<InventoryRecordSet> promise = Promise.promise();
-        InventoryStorage.lookupSingleInventoryRecordSet(okapiClient, instanceQuery).onComplete( recordSet -> {
-            if (recordSet.succeeded()) {
-                JsonObject existingInventoryRecordSetJson = recordSet.result();
-                if (existingInventoryRecordSetJson != null) {
-                    promise.complete(InventoryRecordSet.makeExistingRecordSet(existingInventoryRecordSetJson));
-                } else
-                {
-                    promise.complete( null );
-                }
-            } else {
-                promise.fail(recordSet.cause().getMessage());
-            }
-        });
-        return promise.future();
-    }
-    public boolean foundExistingRecordSet () {
-        return existingSet != null;
-    }
-
-    public boolean gotUpdatingRecordSet () {
-        return updatingSet != null;
-    }
-
-    protected boolean foundSecondaryExistingSet () {
-        return (secondaryExistingSet != null);
-    }
-    public abstract Future<Void> doInventoryDelete(OkapiClient client);
-    // DELETION
-
     public abstract RequestValidation validateIncomingRecordSet (JsonObject inventoryRecordSet);
 
     /**
@@ -292,53 +233,8 @@ public abstract class UpdatePlan {
       }
     }
 
-    public Instance getUpdatingInstance() {
-        return gotUpdatingRecordSet() ? updatingSet.getInstance() : null;
-    }
-
-    public Instance getExistingInstance() {
-        return foundExistingRecordSet() ? existingSet.getInstance() : null;
-    }
-
-    public InventoryRecordSet getUpdatingRecordSet () {
-        return updatingSet;
-    }
-
-    public InventoryRecordSet getExistingRecordSet () {
-        return existingSet;
-    }
-
-    public List<Item> itemsToDelete () {
-        return foundExistingRecordSet() ? existingSet.getItemsByTransactionType(Transaction.DELETE) : new ArrayList<>();
-    }
-
-    public List<Item> itemsToSilentlyUpdate () {
-      return foundExistingRecordSet() ? existingSet.getItemsForSilentUpdate() : new ArrayList<>();
-    }
 
 
-
-    public List<HoldingsRecord> holdingsToDelete () {
-        return foundExistingRecordSet() ? existingSet.getHoldingsRecordsByTransactionType(Transaction.DELETE) : new ArrayList<>();
-    }
-
-    public List<HoldingsRecord> holdingsRecordsToSilentlyUpdate () {
-      return foundExistingRecordSet() ? existingSet.getHoldingsRecordsForSilentUpdate() : new ArrayList<>();
-    }
-
-
-  public List<InstanceToInstanceRelation> instanceRelationsToDelete() {
-        return foundExistingRecordSet() ? existingSet.getInstanceRelationsByTransactionType(Transaction.DELETE) : new ArrayList<>();
-    }
-
-    public boolean isInstanceUpdating () {
-        return gotUpdatingRecordSet() && updatingSet.getInstance().getTransaction() == Transaction.UPDATE;
-    }
-
-
-    public boolean isInstanceDeleting () {
-        return foundExistingRecordSet() && existingSet.getInstance().getTransaction() == Transaction.DELETE;
-    }
 
 
     /* UPDATE METHODS */
@@ -469,90 +365,7 @@ public abstract class UpdatePlan {
 
     }
 
-    /**
-     * Perform instance and holdings updates
-     * @param okapiClient client for inventory storage requests
-     */
-    public Future<Void> handleSingleInstanceUpdate(OkapiClient okapiClient) {
-        Promise<Void> promise = Promise.promise();
-        @SuppressWarnings("rawtypes")
-        List<Future> instanceAndHoldingsFutures = new ArrayList<>();
-        if (isInstanceUpdating()) {
-            instanceAndHoldingsFutures.add(InventoryStorage.putInventoryRecord(okapiClient, getUpdatingInstance()));
-        }
-        CompositeFuture.join(instanceAndHoldingsFutures).onComplete ( allDone -> {
-            if (allDone.succeeded()) {
-                promise.complete();
-            } else {
-                promise.fail(allDone.cause().getMessage());
-            }
-        });
 
-        return promise.future();
-    }
-
-    /**
-     * Perform deletions of any relations to other instances and
-     * deletions of items, holdings records, instance (if any and in that order)
-     */
-    @SuppressWarnings("rawtypes")
-    public Future<Void> handleSingleSetDelete(OkapiClient okapiClient) {
-        Promise<Void> promise = Promise.promise();
-        List<Future> deleteRelationsDeleteItems = new ArrayList<>();
-        for (InstanceToInstanceRelation relation : instanceRelationsToDelete()) {
-            deleteRelationsDeleteItems.add(InventoryStorage.deleteInventoryRecord(okapiClient,relation));
-        }
-        for (Item item : itemsToDelete()) {
-            deleteRelationsDeleteItems.add(InventoryStorage.deleteInventoryRecord(okapiClient, item));
-        }
-        for (Item item : itemsToSilentlyUpdate()) {
-            deleteRelationsDeleteItems.add(InventoryStorage.putInventoryRecordOutcomeLess(okapiClient, item));
-        }
-
-        CompositeFuture.join(deleteRelationsDeleteItems).onComplete ( allRelationshipsDoneAllItemsDone -> {
-            if (allRelationshipsDoneAllItemsDone.succeeded()) {
-                List<Future> deleteHoldingsRecords = new ArrayList<>();
-                for (HoldingsRecord holdingsRecord : holdingsToDelete()) {
-                    deleteHoldingsRecords.add(InventoryStorage.deleteInventoryRecord(okapiClient, holdingsRecord));
-                }
-                for (HoldingsRecord holdingsRecord : holdingsRecordsToSilentlyUpdate()) {
-                  deleteHoldingsRecords.add(InventoryStorage.putInventoryRecordOutcomeLess(okapiClient, holdingsRecord));
-                }
-                CompositeFuture.join(deleteHoldingsRecords).onComplete( allHoldingsDone -> {
-                    if (allHoldingsDone.succeeded()) {
-                        if (isInstanceDeleting()) {
-                          if (getExistingInstance().skipped()) {
-                            InventoryStorage.putInventoryRecordOutcomeLess(okapiClient, getExistingInstance()).onComplete(
-                                handler -> {
-                                  if (handler.succeeded()) {
-                                    promise.complete();
-                                  } else {
-                                    promise.fail(handler.cause().getMessage());
-                                  }
-                                }
-                            );
-                          } else {
-                            InventoryStorage.deleteInventoryRecord(okapiClient, getExistingRecordSet().getInstance()).onComplete( handler -> {
-                                if (handler.succeeded()) {
-                                    promise.complete();
-                                } else {
-                                    promise.fail(handler.cause().getMessage());
-                                }
-                            });}
-                        } else {
-                            promise.complete();
-                        }
-                    } else {
-                        promise.fail(allHoldingsDone.cause().getMessage());
-                    }
-
-                });
-            } else {
-                promise.fail(allRelationshipsDoneAllItemsDone.cause().getMessage());
-            }
-        });
-        return promise.future();
-    }
 
     /* END OF UPDATE METHODS */
 
@@ -604,10 +417,6 @@ public abstract class UpdatePlan {
         if (!repository.getPairsOfRecordSets().isEmpty() && repository.getPairsOfRecordSets().get(
                 0).hasExistingRecordSet()) {
             InventoryRecordSet existingSet = repository.getPairsOfRecordSets().get(0).getExistingRecordSet();
-            if (existingSet.getInstance().isDeleting()) {
-                InventoryRecord record = existingSet.getInstance();
-                metrics.entity(record.entityType()).transaction(record.getTransaction()).outcomes.increment(record.getOutcome());
-            }
             List<InventoryRecord> holdingsRecordsAndItemsInExistingSet = Stream.of(existingSet.getHoldingsRecords(),
                     existingSet.getItems()).flatMap(Collection::stream).collect(Collectors.toList());
 
@@ -637,60 +446,5 @@ public abstract class UpdatePlan {
         return metrics;
     }
 
-    public JsonObject getUpdateStats () {
-        UpdateMetrics metrics = new UpdateMetrics();
-
-        if (gotUpdatingRecordSet()) {
-            metrics.entity(Entity.INSTANCE).transaction(getUpdatingInstance().getTransaction()).outcomes.increment(getUpdatingInstance().getOutcome());
-        }
-
-        if (foundExistingRecordSet()) {
-            if (existingSet.getInstance().isDeleting()) {
-                InventoryRecord record = existingSet.getInstance();
-                metrics.entity(record.entityType()).transaction(record.getTransaction()).outcomes.increment(record.getOutcome());
-            }
-            List<InventoryRecord> holdingsRecordsAndItemsInExistingSet = Stream.of(
-              existingSet.getHoldingsRecords(),
-              existingSet.getItems()
-            ).flatMap(Collection::stream).collect(Collectors.toList());
-
-            for (InventoryRecord record : holdingsRecordsAndItemsInExistingSet) {
-                if (record.isDeleting()) {
-                  metrics.entity(record.entityType()).transaction(record.getTransaction()).outcomes.increment(record.getOutcome());
-                }
-            }
-            if (! existingSet.getInstanceToInstanceRelations().isEmpty()) {
-                for ( InstanceToInstanceRelation record : existingSet.getInstanceToInstanceRelations() ) {
-                    logger.debug("Record: " + record.asJson().encode());
-                    logger.debug("Transaction: " + record.getTransaction());
-                    logger.debug("Entity type: " + record.entityType());
-                    if ( !record.getTransaction().equals( InventoryRecord.Transaction.NONE ) ) {
-                        metrics.entity( record.entityType() ).transaction( record.getTransaction() ).outcomes.increment(
-                                record.getOutcome() );
-                    }
-                }
-            }
-        }
-
-        if (foundSecondaryExistingSet()) {
-            InventoryRecord instanceRecord = secondaryExistingSet.getInstance();
-            metrics.entity(instanceRecord.entityType()).transaction(instanceRecord.getTransaction()).outcomes.increment(instanceRecord.getOutcome());
-            List<InventoryRecord> holdingsRecordsAndItemsInSecondaryExistingSet = Stream.of(
-                    secondaryExistingSet.getHoldingsRecords(),
-                    secondaryExistingSet.getItems()
-            ).flatMap( Collection::stream ).collect( Collectors.toList());
-            for (InventoryRecord record : holdingsRecordsAndItemsInSecondaryExistingSet) {
-                if (record.isDeleting()) {
-                    metrics.entity(record.entityType()).transaction(record.getTransaction()).outcomes.increment(record.getOutcome());
-                }
-            }
-        }
-        return metrics.asJson();
-    }
-
-    public JsonArray getErrors () {
-        // todo: combine error sets?
-        return isDeletion ? getExistingRecordSet().getErrors() : getUpdatingRecordSet().getErrors();
-    }
 
 }
